+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added examples and tests for `isAutoFocusEnabled` prop for `terra-data-grid`.
+  
 * Changed
   * Updated `date-time-picker` KeyboardShortcuts doc for +/- Keys for DateTime input.